[M7] grid: the Tier-0 cross grid — five architectures, 0 divergent
The AOT path and the grid arrive (D-042). Flow: the entry func.func
renames to @frk_entry pre-lowering (the C shim owns main; corpus
protocol: entry symbols are externally-invoked-only), the strategy
pipeline lowers as ever, mlir-translate emits LLVM IR, the PINNED
llvm-22 clang compiles IR to a per-triple object (IR-version safety
against zig's bundled LLVM), and scripts/zigcc.sh (zig cc,
ZIG_VERSION=0.16.0, plain-zig and anyzig shims both handled) links it
with a generated printf shim and the C runtime mirror against
bundled musl/wasi libc — static binaries qemu-user runs sysroot-free.

The C runtime mirror (crates/frk-rt/c/frk_rt.c) is the grid's
runtime: compiled per triple by zig, zero rustup-target setup; the
Rust crate stays canonical for the JIT; the two are held behaviorally
equal BY the grid (aot byte-matches jit on every golden, law L3).

One ABI bug found by the grid's first wasm run: allocator sizes were
size_t in the C mirror while the lowering passes i64 — wasm enforces
exact import signatures and trapped with signature_mismatch. The frk
ABI now says u64 sizes on EVERY target; both twins updated (D-042).

Results — every golden × both strategies, byte-exact:
  x86_64-linux-musl    37/37  37/37   (native)
  aarch64-linux-musl   37/37  37/37   (qemu-aarch64)
  riscv64-linux-musl   37/37  37/37   (qemu-riscv64)
  wasm32-wasi          37/37  37/37   (wasmtime)
  s390x-linux-musl     37/37  37/37   (qemu-s390x, BIG-ENDIAN canary)
The slot model's same-width load/store endian symmetry is now proven
on hardware semantics, not argued from the armchair.

Surface: frnksh grid [--canary|--native] + make grid / grid-native /
canary; ci.sh appends the native slice (continuous compiled-path L3
without a 6-runner dev loop). setup.sh doctors mlir-translate, zig,
qemu-user, wasmtime; versions.env pins ZIG_VERSION and records
wasmtime.

+++ crates/frk-rt/src/lib.rs
@@ -5,8 +5,8 @@
 
 use std::alloc::{Layout, alloc, dealloc};
 
-fn raw_alloc(bytes: usize) -> *mut u8 {
-    let layout = match Layout::from_size_align(bytes.max(1), 8) {
+fn raw_alloc(bytes: u64) -> *mut u8 {
+    let layout = match Layout::from_size_align((bytes.max(1)) as usize, 8) {
         Ok(layout) => layout,
         Err(_) => return std::ptr::null_mut(),
     };
@@ -16,16 +16,18 @@ fn raw_alloc(bytes: usize) -> *mut u8 {
 /// Arena strategy (D-041): bump allocation with process lifetime — the
 /// v0 arena is never reset (region reset entry points arrive with real
 /// region inference). 8-aligned; zero-byte requests return a valid
-/// unique pointer; null only if the host allocator fails.
+/// unique pointer; null only if the host allocator fails. Sizes are
+/// u64 on every target (D-042 — 32-bit-word targets enforce exact
+/// signatures; the c mirror matches).
 #[unsafe(no_mangle)]
-pub extern "C" fn frk_rt_arena_alloc(bytes: usize) -> *mut u8 {
+pub extern "C" fn frk_rt_arena_alloc(bytes: u64) -> *mut u8 {
     raw_alloc(bytes)
 }
 
 /// Rc strategy (D-041): an i64 refcount header sits at `ptr - 8`; the
 /// returned pointer addresses the payload. The count starts at 1.
 #[unsafe(no_mangle)]
-pub extern "C" fn frk_rt_rc_alloc(payload_bytes: usize) -> *mut u8 {
+pub extern "C" fn frk_rt_rc_alloc(payload_bytes: u64) -> *mut u8 {
     let total = payload_bytes.max(1).checked_add(8);
     let Some(total) = total else {
         return std::ptr::null_mut();
